third_party/WebKit/LayoutTests/editing/execCommand/switch-list-type-with-inner-list.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<ul>',
                '<li>^hello</li>',
                '<ol>',
                    '<li>world</li>',
                '</ol>',
                '<li>webkit|</li>',
            '</ul>',
        '</div>',
    ].join(''),
    'insertOrderedList',
    [
        '<div contenteditable>',
            '<ol>',
                '<li>^hello</li>',
                '<ol>',
                    '<li>world</li>',
                '</ol>',
                '<li>webkit|</li>',
            '</ol>',
        '</div>',
    ].join('')),
    'switching an unordered list with a nested list to an ordered list.');
</script>
